Simplify the files
